Moved functions into FoodHotelRegistrationCalc.h
