+++ include/FoodHotelRegistrationCalc.h
@@ -1,0 +1,5 @@
+double calRegCost();
+double hotelCost();
+double *calBreakfast(double amount);
+double *calLunch(double amount);
+double *calDinner(double amount);
